Fix RC update
