+++ packages/amplify_core/lib/src/util/uuid.dart
@@ -16,7 +16,7 @@
 import 'package:uuid/uuid.dart';
 
 class UUID {
-  static const _internal = Uuid();
+  static final _internal = Uuid();
 
   static String getUUID() {
     return _internal.v4();
