+++ packages/amplify_flutter/lib/src/amplify_impl.dart
@@ -22,7 +22,7 @@ import 'package:amplify_auth_plugin_interface/amplify_auth_plugin_interface.dart
 import 'package:amplify_core/amplify_core.dart';
 import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
 import 'package:amplify_flutter/src/config/amplify_config.dart';
-import 'package:amplify_flutter/utils/parse_json_config.dart';
+import 'package:amplify_flutter/src/utils/parse_json_config.dart';
 import 'package:amplify_storage_plugin_interface/amplify_storage_plugin_interface.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
