+++ packages/amplify_flutter/lib/src/utils/create_config_object.dart
