+++ packages/amplify_flutter/lib/src/utils/parse_json_config.dart
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -13,4 +13,18 @@
  * permissions and limitations under the License.
  */
 
-export 'safe_print.dart';
+import 'package:amplify_flutter/src/config/amplify_config.dart';
+import 'package:flutter/foundation.dart';
+
+import 'create_config_object.dart';
+
+AmplifyConfig? parseConfigJson(String configuration) {
+  try {
+    return createConfigObject(configuration);
+  } on Exception {
+    if (kDebugMode) {
+      debugPrint(
+          'There was an unexpected problem parsing the amplifyconfiguration.dart file.');
+    }
+  }
+}
